Test classes moved to resemble real package structure
