+++ web-ui/src/test/java/org/archcnl/domain/TestUtils.java
@@ -1,4 +1,4 @@
-package org.archcnl.domain.input;
+package org.archcnl.domain;
 
 import java.util.LinkedList;
 import java.util.List;
