+++ web-ui/src/test/java/org/archcnl/domain/input/io/AdocEndToEndTest.java
@@ -7,7 +7,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import org.apache.commons.io.FileUtils;
-import org.archcnl.domain.input.TestUtils;
+import org.archcnl.domain.TestUtils;
 import org.archcnl.domain.input.exceptions.ConceptDoesNotExistException;
 import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.model.RulesConceptsAndRelations;
